Predictors and outcomes for people underwent laryngectomy

A dataset containing demographics and surgery details and outcomes on 123 people who underwent laryngectomy at an unspecified medical center between 2000 and 2012.

Usage

laryngectomy

Format

A tibble with 123 rows and 32 variables:

age

Age in years with 90+ removed

gender

sex (M=1, F=2)

recur_site

Recurrent Site (Supra=1, Glottic=2, Transglottic=3, Unknown=4)

trt_mod

Treatment Modality (XRT=1 Chemo+XRT=2 Unspecified=3)

int_stg

Initial Stage (T1=1, T2=2, T3=3, T4=4 Unknown=5)

recur_stg

Recurrent staging (T1=1, T2=2, T3=3, T4=4, unknown=5)

preop_Img_mod

Imaging (CT=1 PET=2 CT+PET=3 MRI=4)

nk_dissection

Neck Dissection (Yes=1, No=2)

Laterality

Laterality (Bilateral Neck Dissection=1, Unilateral Neck Dissection=2)

mic_vas_flap

Microvascular Flap (Yes=1, No=2)

pec_flap

Pectoralis major myocutaneous flap (Yes=1, No=2)

pathology

Nodal Pathology evaluation of specimen (Positive=1, Negative=2)

path_stg

Pathologic staging (T1=1, T2=2, T3=3, T4=4, unknown=5)

complications

Complications (Yes=1, No=2)

fistula

Salivary fistula or leak (Yes=1, No=2)

wound_inf

Wound Infection (Yes=1, No=2)

dehiscence

Wound dehiscence (Yes=1, No=2)

wound_comp

Wound Complications (Yes=1, No=2)

chyle_leak

Chyle leak (Yes=1, No=2)

hematoma

Hematoma (Yes=1, No=2)

revision_proc

Revision procedure (Yes=1, No=2)

flap_fail

Flap Failure (Yes=1, No=2)

periop_death

Perioperative Death (Yes=1, No=2)

med_comp

Medical Complications (Yes=1, No=2)

comp_2plus

2 or more complications (Yes=1, No=2)

hosp_stay

Hospital Stay (<7 days=1, 7-14 days=2, >14 days)

death

Death status (Dead=1 Alive=2)

recur

Recurrence status (Recurred=1, No recurrence=2)

sor

Site of Recurrence after salvage laryngectomy (Locoregional=1, Distant=3, No recurrence=0)

length_fu

Days of follow-up

death_fu

Days until death

recur_fu

Days until recurrence